+++ src/main/java/org/example/BuildingMaterialsLocation.java
@@ -3,7 +3,7 @@ package org.example;
 public class BuildingMaterialsLocation extends BuildingMaterials{
     Coordinates position;
 
-    public BuildingMaterialsLocation(String name,int extractionTime, int constructionValue, Coordinates position){
+    public BuildingMaterialsLocation(String name,long extractionTime, int constructionValue, Coordinates position){
         super(name, extractionTime, constructionValue);
         this.position = position;
     }
